Recreate conditions to show scrollbar bug
If the input field is centered and showing the dropdown elongates the page
enough to cause the browser to spawn a scroll bar, the dropdown gets out of
line (horizontally) with the input.

To reproduce this error:

1. Open `demo.html` in a browser that uses a scrollbar that takes up width
(some newer OSX apps, use a scollbar that hovers above the content).

2. Adjust the browser window height so that it is as short as possible without
having a scrollbar.

3. Type in the query "pa".